- Removed the default order by name for video sources - added a sorting function that sorts by name - added an order field to video sources
