+++ src/db/db.ts
@@ -1,5 +1,5 @@
 import Dexie, { Table } from 'dexie';
-import { VideoSource, initialVideoSources } from '../app/services/video-sources.service';
+import { sortByName, initialVideoSources, VideoSource } from '../app/utils/video-source-utils';
 
 export const dbName = 'watchlist';
 
@@ -9,11 +9,15 @@ export class AppDB extends Dexie {
   constructor() {
     super(dbName);
 
+    const sortedVideoSources: VideoSource[] = initialVideoSources
+      .sort(sortByName)
+      .map((source, i) => {return {...source, order: i}});
+
     this.version(1).stores({
       videoSources: '++id, &name',
     });
 
-    this.on('populate', () => db.videoSources.bulkAdd(initialVideoSources));
+    this.on('populate', () => db.videoSources.bulkAdd(sortedVideoSources));
   }
 }
